requirements.txt
openai==0.27.2
opencv-python==4.7.0.72
urllib3==1.26.11
streamlit==1.20.0
protobuf==3.20.3
python-dotenv==1.0.0
SpeechRecognition==2.2.0